requirements.txt
Django==1.10
django-tastypie==0.13.3
feedparser==5.2.1
pkg-resources==0.0.0
py==1.4.31
pytest==3.0.1
pytest-django==3.0.0
python-dateutil==2.5.3
python-mimeparse==1.5.2
pytz==2016.6.1
six==1.10.0
